+++ recbole/model/layers.py
@@ -924,7 +924,7 @@ class ContextSeqEmbAbstractLayer(nn.Module):
         for type in self.types:
             if len(self.token_field_dims[type]) > 0:
                 self.token_field_offsets[type] = np.array(
-                    (0, *np.cumsum(self.token_field_dims[type])[:-1]), dtype=np.long
+                    (0, *np.cumsum(self.token_field_dims[type])[:-1]), dtype=int
                 )
                 self.token_embedding_table[type] = FMEmbedding(
                     self.token_field_dims[type],
@@ -933,7 +933,7 @@ class ContextSeqEmbAbstractLayer(nn.Module):
                 ).to(self.device)
             if len(self.float_field_dims[type]) > 0:
                 self.float_field_offsets[type] = np.array(
-                    (0, *np.cumsum(self.float_field_dims[type])[:-1]), dtype=np.long
+                    (0, *np.cumsum(self.float_field_dims[type])[:-1]), dtype=int
                 )
                 self.float_embedding_table[type] = FLEmbedding(
                     self.float_field_dims[type],
@@ -1401,14 +1401,14 @@ class FMFirstOrderLinear(nn.Module):
 
         if len(self.token_field_dims) > 0:
             self.token_field_offsets = np.array(
-                (0, *np.cumsum(self.token_field_dims)[:-1]), dtype=np.long
+                (0, *np.cumsum(self.token_field_dims)[:-1]), dtype=int
             )
             self.token_embedding_table = FMEmbedding(
                 self.token_field_dims, self.token_field_offsets, output_dim
             )
         if len(self.float_field_dims) > 0:
             self.float_field_offsets = np.array(
-                (0, *np.cumsum(self.float_field_dims)[:-1]), dtype=np.long
+                (0, *np.cumsum(self.float_field_dims)[:-1]), dtype=int
             )
             self.float_embedding_table = FLEmbedding(
                 self.float_field_dims, self.float_field_offsets, output_dim
